requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
gunicorn==22.0.0
Markdown==3.7
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
